move test_differential_comoving_volume
